update component names
