--- apps/web/src/pages/Auth/LoginPage/LoginPage.stories.tsx
@@ -1,13 +1,0 @@
-import React from "react";
-import { Meta, Story } from "@storybook/react";
-
-import LoginPage from "./LoginPage";
-
-export default {
-  component: LoginPage,
-  title: "Pages/Login Page",
-} as Meta;
-
-const TemplateLoginPage: Story = () => <LoginPage />;
-
-export const Default = TemplateLoginPage.bind({});
