+++ apps/web/src/pages/Auth/SignInPage/SignInPage.stories.tsx
@@ -1,0 +1,13 @@
+import React from "react";
+import { Meta, StoryFn } from "@storybook/react";
+
+import SignInPage from "./SignInPage";
+
+export default {
+  component: SignInPage,
+  title: "Pages/Sign In Page",
+} as Meta;
+
+const Template: StoryFn = () => <SignInPage />;
+
+export const Default = Template.bind({});
