+++ apps/web/src/pages/Auth/SignInPage/SignInPage.tsx
@@ -3,15 +3,15 @@ import { useIntl } from "react-intl";
 import { useSearchParams } from "react-router-dom";
 
 import { Link } from "@gc-digital-talent/ui";
-import { getLocale, useLocale } from "@gc-digital-talent/i18n";
 import { useApiRoutes } from "@gc-digital-talent/auth";
+import { getLocale } from "@gc-digital-talent/i18n";
 import { useFeatureFlags } from "@gc-digital-talent/env";
 import { useTheme } from "@gc-digital-talent/theme";
 
+import Hero from "~/components/Hero/Hero";
+import SEO from "~/components/SEO/SEO";
 import useRoutes from "~/hooks/useRoutes";
 import useBreadcrumbs from "~/hooks/useBreadcrumbs";
-import SEO from "~/components/SEO/SEO";
-import Hero from "~/components/Hero/Hero";
 
 const buildExternalLink = (path: string, chunks: React.ReactNode) => (
   <Link external href={path}>
@@ -19,9 +19,8 @@ const buildExternalLink = (path: string, chunks: React.ReactNode) => (
   </Link>
 );
 
-const RegisterPage = () => {
+const SignInPage = () => {
   const intl = useIntl();
-  const localeState = useLocale();
   const paths = useRoutes();
   const apiPaths = useApiRoutes();
   const { applicantDashboard } = useFeatureFlags();
@@ -36,15 +35,15 @@ const RegisterPage = () => {
   const loginPath = apiPaths.login(fromPath ?? fallbackPath, getLocale(intl));
 
   const pageTitle = intl.formatMessage({
-    defaultMessage: "Sign up using GCKey",
-    id: "e8SFXx",
-    description: "Page title for the registration page for applicant profiles",
+    defaultMessage: "Sign in using GCKey",
+    id: "vkAhEM",
+    description: "Page title for the sign in page for applicant profiles",
   });
 
   const crumbs = useBreadcrumbs([
     {
       label: pageTitle,
-      url: paths.register(),
+      url: paths.login(),
     },
   ]);
 
@@ -102,40 +101,29 @@ const RegisterPage = () => {
               <p>
                 {intl.formatMessage({
                   defaultMessage:
-                    "In order to get started on your profile with the IT Apprenticeship Program for Indigenous Peoples, you <strong>first have to sign up for the platform using GCKey</strong>.",
-                  id: "6TKE1K",
+                    "You can log into your IT Apprenticeship Program for Indigenous Peoples profile using your existing GCKey, even if you've never used this platform before.",
+                  id: "0UeJoO",
                   description:
-                    "Instructions on how to register with GCKey - IAP variant",
+                    "Instructions on how to sign in with GCKey - IAP variant",
+                })}
+              </p>
+              <p data-h2-margin="base(x.5, 0, 0, 0)">
+                {intl.formatMessage({
+                  defaultMessage:
+                    "If you're unsure of whether you have an existing GCKey account, continue to the website and try signing in. If you can't remember your password, you can also reset it there.",
+                  id: "Z+gi2d",
+                  description:
+                    "Instructions on what to do if user doesn't know if they have a GCKey - IAP variant",
                 })}
               </p>
               <p data-h2-margin="base(x.5, 0, 0, 0)">
                 {intl.formatMessage(
                   {
                     defaultMessage:
-                      "GCKey is a service that allows you to sign into <a>many different government services</a> with a single username and password. The button below will take you to GCKey's website, where you'll be able to create an account if you don't already have one. Once you've signed up for GCKey, you'll be brought back here to complete your profile.",
-                    id: "B/xPdU",
+                      "<strong>Don't have a GCKey account?</strong> <a>Sign up for one</a>.",
+                    id: "QsWizb",
                     description:
-                      "Instructions on how to register with GCKey paragraph 2- IAP variant",
-                  },
-                  {
-                    a: (chunks: React.ReactNode) =>
-                      buildExternalLink(
-                        localeState.locale === "en"
-                          ? "https://clegc-gckey.gc.ca/j/eng/ES-01?ReqID=s232921cb5fd612440c0bcab640d5432d02168eb26"
-                          : "https://clegc-gckey.gc.ca/j/fra/ES-01?ReqID=s232921cb5fd612440c0bcab640d5432d02168eb26",
-                        chunks,
-                      ),
-                  },
-                )}
-              </p>
-              <p data-h2-margin="base(x.5, 0, 0, 0)">
-                {intl.formatMessage(
-                  {
-                    defaultMessage:
-                      "<strong>Already have a GCKey account?</strong> <a>Sign in instead</a>.",
-                    id: "7CtKdX",
-                    description:
-                      "Instruction on what to do if user already has a GCKey",
+                      "Instruction on what to do if user does not have a GCKey",
                   },
                   {
                     a: (chunks: React.ReactNode) =>
@@ -181,20 +169,21 @@ const RegisterPage = () => {
             data-h2-justify-content="base(space-between)"
           >
             <p>
-              <Link href={loginPath} mode="inline" external>
+              <Link href={!iapMode ? paths.home() : paths.iap()} mode="inline">
                 {intl.formatMessage({
-                  defaultMessage: "Sign in instead",
-                  id: "Ovlh3a",
-                  description: "Sign in link text on the registration page.",
+                  defaultMessage: "Cancel",
+                  id: "OT0QP3",
+                  description:
+                    "Link text to cancel logging in and return to talent search home.",
                 })}
               </Link>
             </p>
             <p>
               <Link href={loginPath} mode="solid" color="primary" external>
                 {intl.formatMessage({
-                  defaultMessage: "Continue to GCKey and sign up",
-                  id: "Nd1bIG",
-                  description: "GCKey sign up link text on the sign up page",
+                  defaultMessage: "Continue to GCKey and sign in",
+                  id: "4sLCWZ",
+                  description: "GCKey sign in link text on the sign in page",
                 })}
               </Link>
             </p>
@@ -205,4 +194,4 @@ const RegisterPage = () => {
   );
 };
 
-export default RegisterPage;
+export default SignInPage;
