+++ apps/web/src/pages/Auth/SignUpPage/SignUpPage.stories.tsx
@@ -1,0 +1,13 @@
+import React from "react";
+import { Meta, StoryFn } from "@storybook/react";
+
+import SignUpPage from "./SignUpPage";
+
+export default {
+  component: SignUpPage,
+  title: "Pages/Sign Up Page",
+} as Meta;
+
+const Template: StoryFn = () => <SignUpPage />;
+
+export const Default = Template.bind({});
